dzien drugi int

+++ src/Main.java
@@ -3,8 +3,9 @@ public class Main {
         System.out.println("witam panie");
         System.out.println("lekcja programowaniea");
 
-        double wylosowana = (int)(Math.random()*10+1);
+        int wylosowana = (int)(Math.random()*10+1);
         System.out.println(wylosowana);
+
 
         //dzien2
         //git init , git config  user.name user.email , git commit -m "jakas nazwa", git remote add origin , git push - u master
